# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Galaxy ERP",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/tasas": {
      "post": {
        "tags": [
          "Administraci\u00f3n y Cat\u00e1logos"
        ],
        "summary": "Registrar Tasa",
        "operationId": "registrar_tasa_api_v1_admin_tasas_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TasaCambioCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TasaCambioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/tasas/actual": {
      "get": {
        "tags": [
          "Administraci\u00f3n y Cat\u00e1logos"
        ],
        "summary": "Obtener Tasa Actual",
        "operationId": "obtener_tasa_actual_api_v1_admin_tasas_actual_get",
        "parameters": [
          {
            "name": "fecha",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fecha"
            }
          },
          {
            "name": "moneda",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "USD",
              "title": "Moneda"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Obtener Tasa Actual Api V1 Admin Tasas Actual Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/tasas/sincronizar-bcv": {
      "post": {
        "tags": [
          "Administraci\u00f3n y Cat\u00e1logos"
        ],
        "summary": "Sincronizar Bcv",
        "operationId": "sincronizar_bcv_api_v1_admin_tasas_sincronizar_bcv_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Sincronizar Bcv Api V1 Admin Tasas Sincronizar Bcv Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/categorias": {
      "post": {
        "tags": [
          "Administraci\u00f3n y Cat\u00e1logos"
        ],
        "summary": "Crear Categoria",
        "operationId": "crear_categoria_api_v1_admin_categorias_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoriaCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoriaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Administraci\u00f3n y Cat\u00e1logos"
        ],
        "summary": "Listar Categorias",
        "operationId": "listar_categorias_api_v1_admin_categorias_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoriaResponse"
                  },
                  "title": "Response Listar Categorias Api V1 Admin Categorias Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/productos": {
      "post": {
        "tags": [
          "Administraci\u00f3n y Cat\u00e1logos"
        ],
        "summary": "Crear Producto",
        "operationId": "crear_producto_api_v1_admin_productos_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Administraci\u00f3n y Cat\u00e1logos"
        ],
        "summary": "Listar Productos",
        "operationId": "listar_productos_api_v1_admin_productos_get",
        "parameters": [
          {
            "name": "categoria_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Categoria Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponse"
                  },
                  "title": "Response Listar Productos Api V1 Admin Productos Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/productos/{producto_id}": {
      "patch": {
        "tags": [
          "Administraci\u00f3n y Cat\u00e1logos"
        ],
        "summary": "Actualizar Producto",
        "operationId": "actualizar_producto_api_v1_admin_productos__producto_id__patch",
        "parameters": [
          {
            "name": "producto_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Producto Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/productos/importar-excel": {
      "post": {
        "tags": [
          "Administraci\u00f3n y Cat\u00e1logos"
        ],
        "summary": "Importar Productos",
        "operationId": "importar_productos_api_v1_admin_productos_importar_excel_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_importar_productos_api_v1_admin_productos_importar_excel_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Importar Productos Api V1 Admin Productos Importar Excel Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/proveedores": {
      "post": {
        "tags": [
          "Administraci\u00f3n y Cat\u00e1logos"
        ],
        "summary": "Crear Proveedor",
        "operationId": "crear_proveedor_api_v1_admin_proveedores_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProveedorCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProveedorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Administraci\u00f3n y Cat\u00e1logos"
        ],
        "summary": "Listar Proveedores",
        "operationId": "listar_proveedores_api_v1_admin_proveedores_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProveedorResponse"
                  },
                  "title": "Response Listar Proveedores Api V1 Admin Proveedores Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/clientes": {
      "post": {
        "tags": [
          "Administraci\u00f3n y Cat\u00e1logos"
        ],
        "summary": "Crear Cliente",
        "operationId": "crear_cliente_api_v1_admin_clientes_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Administraci\u00f3n y Cat\u00e1logos"
        ],
        "summary": "Listar Clientes",
        "operationId": "listar_clientes_api_v1_admin_clientes_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClienteResponse"
                  },
                  "title": "Response Listar Clientes Api V1 Admin Clientes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/zonas": {
      "post": {
        "tags": [
          "Administraci\u00f3n y Cat\u00e1logos"
        ],
        "summary": "Crear Zona",
        "operationId": "crear_zona_api_v1_admin_zonas_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZonaCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZonaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Administraci\u00f3n y Cat\u00e1logos"
        ],
        "summary": "Listar Zonas",
        "operationId": "listar_zonas_api_v1_admin_zonas_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ZonaResponse"
                  },
                  "title": "Response Listar Zonas Api V1 Admin Zonas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_importar_productos_api_v1_admin_productos_importar_excel_post": {
        "properties": {
          "archivo": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Archivo"
          }
        },
        "type": "object",
        "required": [
          "archivo"
        ],
        "title": "Body_importar_productos_api_v1_admin_productos_importar_excel_post"
      },
...
      "CategoriaCreate": {
        "properties": {
          "codigo": {
            "type": "string",
            "maxLength": 30,
            "minLength": 1,
            "title": "Codigo"
          },
          "nombre": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Nombre"
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          }
        },
        "type": "object",
        "required": [
          "codigo",
          "nombre"
        ],
        "title": "CategoriaCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TasaCambioCreate": {
        "properties": {
          "fecha": {
            "type": "string",
            "format": "date",
            "title": "Fecha"
          },
          "moneda": {
            "type": "string",
            "maxLength": 3,
            "title": "Moneda",
            "default": "USD"
          },
          "fuente": {
            "type": "string",
            "maxLength": 20,
            "title": "Fuente",
            "default": "MANUAL"
          },
          "valor": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string"
              }
            ],
            "title": "Valor"
          }
        },
        "type": "object",
        "required": [
          "fecha",
          "valor"
        ],
        "title": "TasaCambioCreate"
      },
      "TasaCambioResponse": {
        "properties": {
          "fecha": {
            "type": "string",
            "format": "date",
            "title": "Fecha"
          },
          "moneda": {
            "type": "string",
            "title": "Moneda"
          },
          "fuente": {
            "type": "string",
            "title": "Fuente"
          },
          "valor": {
            "type": "string",
            "title": "Valor"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "fecha",
          "moneda",
          "fuente",
          "valor",
          "created_at"
        ],
        "title": "TasaCambioResponse"
      },
...
    }
  }
}